move ipc helpers to compass-utils
