+++ packages/compass-utils/package.json
@@ -51,7 +51,8 @@
     "reformat": "npm run prettier -- --write . && npm run eslint . --fix"
   },
   "optionalDependencies": {
-    "@electron/remote": "^2.0.10"
+    "@electron/remote": "^2.0.10",
+    "electron": "^23.3.10"
   },
   "devDependencies": {
     "@mongodb-js/eslint-config-compass": "^1.0.6",
